Refactored code using Various Refactoring Methods to Improve Code Quality.
