+++ core/src/main/java/com/digitalpebble/stormcrawler/filtering/basic/CharacterEscaper.java
@@ -1,0 +1,5 @@
+package com.digitalpebble.stormcrawler.filtering.basic;
+
+public interface CharacterEscaper {
+    String escape(byte b);
+}
