+++ core/src/main/java/com/digitalpebble/stormcrawler/filtering/basic/NonControlCharacterEscaper.java
@@ -1,0 +1,8 @@
+package com.digitalpebble.stormcrawler.filtering.basic;
+
+public class NonControlCharacterEscaper implements CharacterEscaper {
+    @Override
+    public String escape(byte b) {
+        return String.valueOf((char) b);
+    }
+}
